[Frontend] Number widget refetch (#14448)
