+++ opencti-platform/opencti-front/src/components/dashboard/DashboardEntitiesViz.tsx
@@ -1,6 +1,6 @@
 import { memo, ReactNode } from 'react';
 import StixDomainObjectBookmarksList from '../../private/components/common/stix_domain_objects/StixDomainObjectBookmarksList';
-import StixCoreObjectsNumber from '@components/common/stix_core_objects/StixCoreObjectsNumber';
+import StixCoreObjectsNumber from '../../private/components/common/stix_core_objects/StixCoreObjectsNumber';
 import StixCoreObjectsList from '@components/common/stix_core_objects/StixCoreObjectsList';
 import StixCoreObjectsDistributionList from '@components/common/stix_core_objects/StixCoreObjectsDistributionList';
 import StixCoreObjectsMultiVerticalBars from '@components/common/stix_core_objects/StixCoreObjectsMultiVerticalBars';
@@ -61,13 +61,13 @@ const DashboardEntitiesViz = ({
         <StixCoreObjectsNumber
           variant={undefined}
           height={undefined}
-          endDate={endDate}
-          startDate={startDate}
           dataSelection={widget.dataSelection}
           entityType={undefined} // because calling js component in ts
           parameters={widget.parameters as object} // because calling js component in ts
           popover={popover}
           host={host}
+          refreshRate={refreshRate}
+          config={config}
         />
       );
     case 'list':
